requirements.txt
fastapi
uvicorn
yt-dlp
python-multipart
python-dotenv
requests
